🙈(gitignore) remove the entry for data
We gitignored `data` because we used to have a data/ folder at
the root of the project where we put static files and build
results.

This is not the case anymore, so we can remove this entry. This
is useful as we're about to create a `data` folder in our frontend
app.