➕(dependencies) add redux & react-redux to our dependencies
Marsha is beginning to outgrow its custom state management using
the `App` component state and a custom context.

We decided to use this opportunity to switch to a standardized
solution (redux, as in richie, our other frontend project). This
should not be too painful as we can reuse what we already built
for richie, and prevents us from building out our homegrown solution
too much.