♻️(frontend) split Dashboard video matters into a separate component
As we're about to add timed text management to the Dashboard, we need
to use a separate DashboardVideoPane component to display the part of
the Dashboard that is about the video and hold the related logic.

To do this, we removed a little "hack" which consisted in having VideoForm
display a Dashboard during upload to avoid making upload states shared
between components. This meant updating the video directly in the store
to reflect its client-side state rather than what the record on the backend
probably holds.

We also move dashboard-related components to their own folders from
components/Dashboard to keep a cleaner structure.