✨(frontend) add a link to the dashboard from the video form
There are some circumstances when users need to be able to browse
from the video form to the dashboard: when they clicked to update
the video but elected not to upload a new file, or when they want
to update timed text files before the video.

To accomodate these screnarii, we added a link in the video form
that points to the dashboard.