♻️(types) group track-related types in a common file
As we're about to add a type for timed text tracks and some
related enums, we decided it would be more tidy to keep those
types together in the same file, mirrorring the organization
of code on the backend side, rather than split them into
an arbitrary number of files.

We also need to make apparent the fact that states are shared between
different types of tracks. Rename `videoState` to `trackState`
so it can be cleanly used for both video tracks and timed text
tracks.