♻️(frontend) make <VideoForm /> & <VideoUploadField /> generic
We need to reuse the <VideoForm /> and <VideoUploadField /> components
for timed text files, and later on for other kinds of files.

Rename them to <UploadForm /> and <UploadField /> respectively, and
make them more generic so we can easily use them for other files than
videos.

This includes using URL parameters and an `objectType` prop to set
them up on load.