🐛(api) fix a test that randomly broke with a 1/6 probability
We wrote a new test to make sure the GET LIST request on the API
for TimedTextTracks works. It creates 2 TimedTextTracks. This causes
it to randomly break if we call the factory without language and
mode parameters because TimedTextTracks are unique for each
language-mode-video combination.

Simply forcing different modes prevents those random fails.