♻️(frontend) fix the link to <UploadForm /> in <InstructorView />
We recently updated the upload component (now callled <UploadForm />)
to be more generic. It now needs to be passed a model name and an
object ID.

We did not update the link in the <InstructorView /> to account for
that change. We need to get the video from context (hence the
`Connected` wrapper) and pass it to <InstructorView /> so it can
generate a working link.