✨(frontend) prepare the app to handle timedtexttracks
Until now marsha only supported one kind of Resource, Video. As
we're adding timed text tracks, we need to add in some relevant
types and adapt our data (RootState & reducer) to handle them.

We also took this opportunity to add typings for schemas &
route OPTIONS as we are going to need them.