♻️(upload) rename state field to upload_state
The name `state` for eponym field on the video and track models will become
confusing in a future refactoring because it will not represent the state of
the object anymore but the state of its latest upload.

The new upload_state field will be reset to `pending` by the client each time
it initiates an new upload. This is necessary because we currently fail to
follow-up on an upload in case of an update to a video/track that has already
been successfully uploaded. We realized that it is better to rely on the
`current_stamp` field to know if a video has already been uploaded instead of
looking at the state field. The state field will now actually track the state
of the latest upload and should thus be renamed to `upload_state` for clarify.