✨(api) add field `is_ready_to_play` to compensate removing field `state`
In the previous commit we removed the `state` field (see explanations in
commit). In order to make it clear and straightforward to check the state
of a video (resp track), we add this convenience field trivially computed
from the value of the active_stamp field. It answers to the question:
"has this object been successfully uploaded before". Indeed it would mean
that the video (resp track) points to a valid file in S3 and can be played
eventhough the upload_state may now be "pending" or "error" because an
update was attempted and has not yet completed sucessfully.