♻️(upload) replace getting an upload policy by initiating an upload
as suggested by @lunika, making the upload_state field updatable is not
optimal and requires the client to make 2 successive request to initiate
an upload: get the upload policy + reset the upload state to "pending".

In this commit as rename the API endpoint to "initiate-upload" and make
it a POST so that we can reset the upload state before issuing the upload
policy to the client.