♻️(frontend) update state to upload_state & is_ready_to_play
We renamed the `state` field on Videos and Tracks to `upload_state`
and added the `is_ready_to_play` field to better manage updates
to existing objects: this allows us to keep using eg. existing
videos while a new version is being uploaded and processed.

To reflect these changes in the backend, we modified the objects in
the frontend and updated some cases where videos should be treated
differently not only based on `upload_state` but also based on
the new `is_ready_to_play` flag.