🐛(video) fix duplicating videos for playlist/site portability
When duplicating a video, we forgot to set the "resource_id" field
so we ended-up with copies that pointed to a file that did not
exist in S3. Furthermore, the video was duplicated endlessly each
time the instructor was trying to access it. Big bug indeed.

It was not spotted because of insufficiant tests so I added the
missing assertions to existing portability tests.