🐛(lambda) fix an issue that prevented proper logging for lambda-encode
We forgot to return the MediaConvert Job Data from `encodeVideo` into
the handler; this causes it to fail when it attempts to log said
Job Data it did not receive.

The problem was not caught in the tests because the whole code block
was wrapped in a `try/catch` block to catch side-effect related issues.

We reduced the scope of the `try/catch` to only encompass side effect
calls and make the relevant unit test fail. We could then simply update
both `encodeVideo` and our mock to have the expected behavior of returning
the Job Data.