♻️(frontend) replace pre-upload GET policy with JIT POST init-upload
We previously issued a GET /policy-upload/ request as the <UploadForm />
component started up. We then used this policy to perform the AWS S3
upload when the user ended up adding a file. If they did not, the
GET operation did not cause any side-effects.

We replaced this with an `/initiate-upload/` that both gets a
policy and resets the objects `upload_state` to PENDING. We therefore
changed it from a GET to a POST to reflect the nature of the request.

This also means we need to perform it as late as possible to avoid
resetting the `upload_state` when the user is not actually adding a
file. Instead of doing it at `componentDidMount` time, we now wait
for the user to add a file before we perform this request. There, we
trade a slightly longer load time for the user, for more reliable
information on the object's upload state.