🚸(ltipassport) enforce either consumer site or playlist field
When creating an LTI passport via the admin, it was not required to
set either a playlist or consumer site but leaving both blank was
making the passport unusable and generating 500 errors.

On the opposite, when setting both a consumer site and a playlist it
was not clear that the consumer_site would take precedence. A better
experience is to forbid setting both so that the passport configuration
is always explicit.

In this commit with add checks upon saving an LTIPassport instance to
ensure that one and only one of the consumer_site or playlist fields is
set. We added tests to secure this behavior.