✅(lti) improve portability test to same playlist on another site
We forgot to test that the video from the same playlist had precedence
over videos from other playlists. In order to test this, we should add
another video linked to another playlist and that is newer than the
video linked to the same playlist. It would be preferred if it wasn't
for the same playlist.