✨(lti) add automatic portability between consumer sites
By default, the portability of videos between consumer sites is not
allowed. It has to be activated by the instructor or an admin at the
level of a playlist (see is_portable_to_consumer_site).

In some cases, we want videos to be automatically marked as portable
to another consumer site:
- portability between different environments of a website (staging/
  preprod/production...)
- portability between friend websites or websites operated by the
  same company, etc.

This commit adds a new many to many field to the consumer site model
allowing to declare a list of other consumer sites to which all the
videos of the consumer site can be transferred automatically without
further declaration of portability at the level of the playlist.