🚨(tslint) remove the max-line-length rule from tslint
The max-line-length rule from `tslint` conflicts in some cases
with the formatting we get from `prettier`.

Our policy is to always prioritize `prettier` as it does the
actual formatting job for us and is a more generally accepted
tool for these cases.

We can plainly remove the `tslint` rule as line-length is
enforced by `prettier` for all cases anyway.