📝(readme) write an actual README.md file for Marsha
As the project is nearing a stable state, being deployed to client
environments and soon to be presented at conferences, it needs to
have a proper README to enable outside developers to install it, use
it and eventually contribute.

Start with a minimally appropriate overview, description of the
architecture and "getting started" tutorial. Those are the most
important parts to make the project more approchable and start
collecting feedback on our documentation.