⬆️(node) use node 10 to run builds & tests
Node 9 is EOL and Node 8 is soon to be, and is already an old
version at this point.

Switch Node versions in Dockerfiles & CircleCI builds to Node 10,
which is the current LTS.

NB: Alas, AWS Lambdas do not currently support more modern versions
than 8.10. We therefore had to ensure tests for them still run on
the same version they will be deployed on.