♻️(development) update the LTI development view
We changed the URL for the LTI Video view. It now needs to include
a UUID for the targeted resource. We therefore need to add a field
in the development view's forms for LTI request params.

As UUIDs are not practical to write by hand for users, we also added
a randomly generated one to the context data to use as a default in
the form.