⬇️(dependencies) keep `react-dropzone` in a lower version for now
As we just added renovate to Marsha, we had a massive pull request
with several dozens of dependency upgrades. Most were not too
painful and we were able to bundle them together.

`react-dropzone` is a little bit trickier so we'll refrain for
now and upgrade it next time we get a dependency PR.