🔧(build) update `Jest` config for updated versions
Update Marsha's testing configuration to the same as Richie's
which works for `Jest` and our build tools' latest versions.

Note: this config update breaks `jest.doMock` as all `imports`
are now hoisted. This means we need the hoisting behavior of
`jest.mock` to be able to actually mock any module.