🔥(front) remove our `styledComponentWithProps` shim
We used the `styledComponentWithProps` shim to enable types and
typechecking in custom styled components' props. This was a bit
unwieldy and broke syntax highlighting in the TS language server.

With the improved types brought by newer versions of
`styled-components`, we can just pass common `Props` to the
template strings functions.

Note: this newer version of `styled-components` also introduced a
different way to add global styled, which we refactored to.