🚑(frontend) hide the <DashboardTimedTextPane />
We need to hide the <DashboardTimedTextPane /> for now to enable
new deployments. We want to avoid confusing users until we are
able to display the subtitles, closed captions & transcripts in
the video player and to make some improvements to the timed text
uploading process.

We can easily add it back as soon as we want by just adding back
<DashboardTimedTextPaneConnected /> to the <Dashboard />.