➕(design) add Grommet to our dependencies & index
The need for a frontend design system and base library for Marsha's
frontend recently arose. We decided on Grommet as a library based
on `styled-components`, accessible, fully typed and "batteries included".

Simply add it to our dependencies and add the base <Grommet /> component
at the root of our component tree in index to enable theming.