🗃(videos) make data migration on resource_id field before removing it
The `resource_id` field is the value used by the old LTI endpoint to find the
video in Amazon S3. The new LTI endpoint uses the `id` field so we must
modify the existing videos and replace their id by their resource_id so that
they continue  to work.

The schema migration to remove the field is pushed back to a future migration
to be run in a separate deployment.