🏷️(grommet) add utils typings for grommet
Some of the `grommet/utils` are very useful to us as consumers
of the library. `normalizeColor` notably, is the way we're supposed
to retrieve colors from a theme.

To be able to use them, we need basic typings for them. Until some
are added to grommet itself, we need to add them to our own types
folder.
See https://github.com/grommet/grommet/pull/2655 for more on this.