🏷️(types) improve the base `Resource` type
The base `Resource` type specified the `ìd` as either a string or
a number. As we're relying on UUIDs, it's useless to allow numbers,
especially as any future non-UUID ids could be treated as strings
anyway.

As we made this change, we noticed `Resource.ts` did not hold
anything meaningful anymore, apart from a duplicate type definition
for the `resourceById` `reducer. We therefore consolidated `Resource`
into `tracks.ts` and got rid of `Resource.ts`.