✨(upload) update the store to process upload progress notifications
Upload progress is inherently shared state, as the component (or
service) that manages the upload is not the same one that is
responsible for informing the user.

We thus decided to put it into the store as part of the `context`
slice of `RootState`. We wrote a reducer for it with an
`ìnitialState`, which prompted us to move `appData` out of the
index to avoid circular dependencies.