✨(upload) have <UploadForm /> dispatch upload progress notifications
As we wanted to be able to monitor upload progress, we had to use
the `XMLHttpRequest` API instead of fetch as fetch does not provide
any progress indicators.

We isolated this part in a side effect function and left
`UploadForm` in charge of managing the process and provide connection
(through `<UploadFormConnected />`).

We could then simply dispatch a redux action whenever the browser
emits some information through the `progress` event on
`XMLHttpRequest`