✨(upload) show current upload progress in <DashboardVideoPane />
We want to give users information on the upload progress on the
`<Dashboard />` so they can be reassured everything is working
correctly. This is necessary because videos can be large files, and,
when uploads take a long time, a user facing a spinner may start
thinking there is an issue and just leave.

We're using `Grommet` supplied components to quickly implement this,
using the progress information from `RootState['context']`.

To improve the experience on the `<Dashboard />`, we also made sure
to hide the "Replace" and "Watch" buttons until the video is actually
ready.