🔧(storage) use manifest static files storage class for deployed envs
The static files (css/js) often change but because the files keep the
same name, a browser that has it in its cache will not get the new
files immediately and will therefore continue to use an outdated version
of such files.

Django proposes a file storage that automatically renames the files to
include a hash of their content in their name. The mapping between the
original file name and the hashed name is stored in a manifest file
(hence the name of the storage backend).