♻️(frontend) replace our <Button /> & colors with Grommet's
We implemented our own <Button /> component and began building a
theme before we adopted Grommet. Now that we did, they are redundant,
which is why we're replacing our <Button /> with Grommet's and
getting rid of our color definitions.

As we worked on the buttons in `<DashboardVideoPaneButtons />`, we
noticed an issue: we only showed the buttons when the video was
`READY`, but we actually need the "Upload" button to let the user
go back to the `<UploadForm />` when the video is `PENDING`. We
took this opportunity to fix that up.