✨(frontend) manage timedtexttrack lists in the reducer
Before we can integrate the related side-effect with redux-saga,
we need to have a reducer ready to handle resource-list related
actions for timedtexttracks.
Add a generic reducer and apply it to the timedtexttracks reducer.

As we did this, we noticed our types were too lax when using the
`Reducer` and `AnyAction` types from redux. We removed them and
made our own global action type to let typescript discriminate
the actions for us.