➕(dependencies) add redux-saga to our dependencies
As marsha grows from a little iframe-baseds utility to a full-fledged
application with different views, interconnected objects and a lot
of different API requests, we need to adopt tools to manage side
effects.

As we're already on redux and we already adopted saga for some of
our other projects, we opted to use it in marsha.