✨(frontend) add a saga to handle resource list GETs
One of our most simple needs when it comes to side effects is to
GET list of resources, and store them both as a list and as a series
of unique resources.

We can implement a simple saga using the `resourceList` reducer's
actions. We also need to set up the base saga plumbing with relation
to the redux store as it is our first saga.