♻️(frontend) use the reducer & saga to get the timedtexttracks
Now that the resource list reducer and corresponding saga are
available, we can refactor <DashboardTimedTextPane /> and its
connected container to use them instead of the previous direct
side effect.

This will be helpful to manage the list when effects happen that
affect one item as opposed to the whole list. It also makes the
component more lightweight.