👷(frontend) run webpack in production mode for deployed builds
We overlooked it until now, but it is absolutely necessary to run
webpack in production mode. Otherwise, a lot of development utilities
in React and other libs are included, and the code is not minified.

To do this for deployed builds, we need to set this mode in the
`Dockerfile` at the root of the repo, as it is the bundle generated
there that is deployed to live environments.

It can easily be overwritten by development builds run directly
through `yarn` on developer machines.

NB: updated yarn.lock to get terser 3.16.1 (dep of a dep) instead
of 3.16.0 which had an issue that prevented the production build
from working.