✨(backends) use S3 to store and Cloudfront to distribute static files
We want to optimize the time it takes for Marsha to load and start.
Downloading the javascript bundle is on the critical path so the faster
it is downloaded, the faster the player will be visible.

Note that we make the manifest file name configurable so that we can
version it with our blue/green deployments and avoid conflicts between
2 versions of Marsha running side by side.