🏷️(tests) use jestMockOf helper to simplify jest.Mock type use
Newer versions of jest types have better typings that actually
extract the parameters and return types for mocks instead of just
using `any` or `unknown`.

They are also pretty unwieldy for our use-case as they explicitly
require the parameters and return values to be passed to them. This is
not practical when we use them directly as we're often just "holding"
the mocked function itself.

Our helper just uses `jest.Mock` but takes the type of the mocked
function as argument instead of the types of its parameters and return
value.