🍏(player) add a feature check for MSE support before using DashJS
The `<VideoPlayer />` did not run a feature check before initializing
dashjs. This caused an exception on iOS devices as they do not offer
the MSE that dashjs relies on. There, ABR is supported through builtin
HLS instead.

We can simply add a check for MSE and only load dashjs when they are
available. iOS can then use HLS and other non-MSE-compatible browsers
will just use MP4 sources.

We took this opportunity to stop loading MP4 sources when dashjs is
enabled as loading the MP4 sources adds a UI to select video definition,
which is always overridden by dashjs.