💩(player) attempt to fix video playback on iOS devices
iOS devices are supposed to manage Adaptive bitrate playback with the
HLS spec natively. For a reason we have yet to uncover, we get errors
as we try to play an HLS video (our current theory is that our manifest
file and/or video fragments may be in a format iOS devices do not
support).

We therefore removed HLS from the sources so iOS devices will just use
regular old `.mp4` files. We cannot be sure yet that this works as our
development environments fail due to CORS issues (that do not exist on
staging & production environments).