🐛(docker) change frontend build to fix translations
To compile translated messages in the translation file, we need the
source extracted by babel-plugin-react-intl. To extract them we need to
build the project and to build the project we need a translation file...
So we fake an empty translation file and then do what we describe before
to have a valid build including translations.