♻️(aws) refactor AWS utilities to use a bin script
We had created too many make commands (one for each Terraform command)
and still we could not pass arguments and run commands that were not
planned. It is more convenient to use a bin/script to run any
Terraform command. I also reverted to use the official Terraform
image as it was done on the openfun/fun-infra project.