🐛(statics) dynamically set webpack publicPath
Webpack uses the property `output.publicPath` to import chunks. If we use
a CDN, and it's our case, the CDN url should be set in the `publicPath`
but it's not possible to know it at build time. To fix this issue
the variable __webpack_public_path__ can be changed at runtime and then
used to fetch chunks. See the documentation:
https://webpack.js.org/configuration/output/#outputpublicpath