⬆️ (backend) update from django 2.1.8 to 2.2.1
We want to use the last django version. With django 2.2 it is not
possible anymore to use django-postgres-extra because Index can now have
a condition and this condition property is not compatible with the one
used by ConditionalUniqueIndex.
To resolve the issue we directly used Index with a condition but now the
created index is not unique anymore.