✅(router) add helper to easily use routes in tests
As we use testing-library's complete rendering in our tests, we
need to wrap our tested components in a `react-router-dom` router
if they use any of its utilities or need to render <Redirect />
or <Link /> components.

As such, this does not warrant adding a utility. However, we sometimes
need to manage and test the redirects, which is not easy to achieve
through mocking. We can instead have our utility mock our redirect targets
and replace them with trivial elements to test them.