📌(AWS) pin terraform to version 0.11.14
We must pin the terraform version we use to version 0.11.14. There are
breaking changes in version 0.12 and we have to do a migration. Moreover,
we also need to migrate the AWS version which also contains breaking
changes. In a first time we will use a version we know and compatible
with our existing configuration, workspaces and states.