♻️(front) reactivate HLS support
We have file to be used by iOS base on the HLS spec but we remove them
in version 2.5.0. We noticed that when both HLS and mp4 are used, the
ios player will not used the HLS file. Here we only used the HLS file if
the browser can play it, we use mp4 file otherwise.