♻️(frontend) refactor TimedTextCreationForm
As part of our ongoing refactoring of marsha's frontend, we're working
on <TimedTextCreationForm /> to update tests to @testing-library, fold
connect-related stuff into the component itself, and move from a
class component to a function component.

This also enables us to use hooks and start using lightweight data
management (eg. zustand) in the component.