♻️(front) refactor UploadForm component
This commit is part of our ongoing refactoring of marsha's frontend. The
UploadForm component is now a functional component and the connection to
redux is made in one module.
Also, is order to easily and fully test the `upload` function we choosed
to extract it out of the component, as a standalone function and fully
test it.