♻️(frontend) move object upload progress tracking to a zustand store
The context reducer in our Redux store only includes static data
that does not need to be included in a store, and upload progress
tracking.

To prepare ourselves to remove the context reducer, we can begin by
tracking objects upload progress through a very simple zustand store.

This also lets us delete the separate connected versions of
<DashboardObjectProgress /> whose only purpose was to connect to the
progress in the context state slice.