♿️(front) set tab-index to -1 for plyr play-large button
Play-large button has the exact same behaviour as main play button in
the control bar. This can be confusing for screen reader and lead to
think that there is two videos instead of one. By setting the tab-index
to -1 the screen reader will ignore it and only detect the main play
button.