🐛(frontend) fix dashjs v3 breaking changes
We recently upgraded dashjs to v3 as part of our regular dependency
upgrade schedule. This new version had some breaking changes which
we did not take into accound and which did not break tests as those
mock the dependency.

Also, this v3 provides its own (imperfect) type definitions, which
means we do not need to roll our own any more.