🐛(frontend) fix an issue with video polling
The Dashboard video pane component is supposed to poll the video
until it is ready, when it renders a non-ready video. This was not
happening when uploading a new video to replace an existing, ready
one, because the dashboard was first called with outdated data
pertaining to the ready video.

We just had to move around the redirection to the dashboard video
pane when the user picks the replacement video, to make sure it
happens after the video object has been updated to its new state.