♻️(front) change architecture to reuse component with documents
Many components can be reused with documents and videos. To do that we
renamed some of them. We also introduce lazy components loaded only when
we need them depending if we are working with a video or a document.
This also changes a little bit the way we manage component. Until now we
didn't have component exported as default.