🐛(frontend) remove  our dependency on the "querystring" module
Marsha was using the "querystring" module without declaring it as a
dependency. This was working because another dependency depended on
it, so, even though we did not declare it, it was always present in
our node_modules.
Typescript was also accepting it as it has the same name as a code
NodeJS module.

Since our use case is so simple here (we only have 2 params), we do
not need to import a generic module for it.
We can simply use a template string and remove our dependency.

We also took this opportunity to improve the tests by checking error
reporting, which also removes unwanted logs from tests.