♻️(back) change video content-disposition header when downloaded
We want to change the filename when a video is downloaded instead of
having the video hash we use to store it in S3. The content-disposition
header response can be changed if the url is signed. We use this
querystring to change the filename with the playlist title and the
video position in this playlist.