✨(back) manage extension for document resource
For the document resource we need to know its extension in order to add it
when we craft the filename that the file will have on the user's filesystem
after downloading it. To do that we guess it from the initial
filename/mimetype, use it in the S3 key and retrieve it in the update
state endpoint.