🔥(front) remove duplicated code in PollForTrack function
A type guard was made in pollForTrack function to determine if a
resource was ready to show. In PR #499 we renamed all is_ready_*
properties in is_ready_to_show, when the refactor was made our editor
changed the name everywhere and we didn't notice that this type guard
was useless.