🐛(front) change how we poll timed text track when we replace them
When a timed text track was already in READY state and is_ready_to_show
set to true, it was blocked in pending state if we use the replace
feature. The issue is in the useEffect of TimedTextListItem component,
we start polling only when the is_ready_to_show property was equal to
false. Doing like this is wrong, we have to check the updload state and
start polling when the state is pending, uploading and processing.