♻️(configuration) remove redundant setting STATICFILES_AWS_ENABLED
Checking if the CLOUDFRONT_DOMAIN setting is set should suffice.
Having both can lead to confusing situations when one is set but
not the other one when in fact they are the same.

This setting had not been added to docs/env.md so we don't need
to update the doc. It is not a breaking change as we are now
checking the CLOUDFRONT_DOMAIN setting that was already set.

Reviewed and fixed by @lunika