♻️(statics) change how static base url is computed
In commit d9c3fd5 we decided to rely on the CLOUDFRONT_DOMAIN setting to
determine which staticfiles_storage is used and then use the `url`
method to compute a base url. In local this work fine but we use the s3
storage class it doesn't work anymore because the staticfiles manifest
doesn't know an empty filename. To fix this, we want the js folder
path, we know it always exists and it is exactly the path we want in
front application. So we compute the url for the js folder and we don't
have to append /js at the end of the base url in the front application.
We can use the value in meta name public-path directly.