🐛(back) compute base static url from an existing file
In commit b22477c we tried to fix how we compute the static base url.
In our tests it was working but we didn't test it in Production
environment. In the environment we tested it, the DEBUG mode was
activated and in this mode the S3Boto3Storage class storage always
return a result even if the requested file does not exist. To fix our
issue we choose to request an existing file: js/index.js
This file is the frontend entrypoint. It will always exist
until the frontend application choose to change it...