✨(AWS) create first migration to escape existing timed text tracks
Escaping timed text tracks is the first migration we need. We want to
run the new lambda encode on all already uploaded timed text tracks. In
a first we list all objects uploaded in the source bucket, we test is
the key is a timed text track and invoke the lambda with this key as
parameter.