🔧(build) manage polyfills for dependencies
Our code is now built, transpiled and polyfilled to work in IE11. This,
however, is not enough to ensure Marsha can even run there. Some of
our dependencies are not built for IE11 and we need to address them.

First, we have a few that need actual transpilation to run, we can just
add them to our test for the babel-loader. This is also true of our own
'public-path' utility.

For others, we need to add polyfills. We were using babel's "usage" setting
to automatically polyfill only language features we actually use. Since we
need to also polyfill for our dependencies, and those do not go through babel,
we need to be more aggressive with our bundle and use "entry".