🐛(back) ignore domain check if http referer is missing
In the LTI.verify method we check if the referer matches the
URL for the consumer_site associated with the passport. In some cases
users block the referer for privacy concern and we can not do the check
and an error is raised. To avoid this error we choose to not verify the
domain if the referer is absent.