♻️(back) set XAPI timestamp in enrich_statement method
We do not rely on the timestamp sent by the front application. This
timestamp can be wrong, modified by the user and so not relevant. We
choose instead to use the server timestamp, it will not be exactly
when the event occured but at leat it will be always relevant.