✨(back) host plyr icons svg
Icons in plyr interface are loaded fron an svg file. This file is hosted
by plyr itself on their own CDN. We want to host it on our own
infrastructure to prevent their CDN issues. For this we copy the svg
file in the static directory and use the django static file storage
system to compute its url and expose it in the app data in the lti view.