🐛(back) fetch resource based on uploaded_on field and not upload_state
The method get_or_create_resource looks for the resource searching with
upload_state="ready". In some cases the resource can be shown when the
upload_state is not ready (pending update upload). For this we already
implemented the uploaded_on field and if not None then the resource can
be shown because we know there is a resource available on S3 and urls
can be computed.