♻️(backend) move defaults settings to settings file
Some defaults settings where declared in the defaults module and looking
in settings if it was declared to fallback then on a default value. This
don't allow us to change these settings with an environment variable and
moreover fallbacking on default value can also be done in settings
module. We decided to move them in the settings module.