🐛(front) choose video quality when ABR is not available
When browser does not support ABR, user should be able to choose the
video quality. It wasn't possible because video sources where not in the
DOM when plyr is instantiated. The only way to have them is to recreate
sources (video source and timed text track) and inject them in plyr
instance.