🐛(xapi) initliaze xapi module only when all video data are available
In some cases fetching all video data can take more time than expected
and some data are missing in the canplay event. Do be sure to have all
needed data we have to use others events (loadedmetadata and loadeddata)
and initialize the module only when the data are avaialble in one of
this 3 events.