🐛(front) set plyr sources also if hls is not available
In commit fe357 we set plyr sources if ABR is not available but not when
hls is also not available. The consequence is that hls is no more used
when available in favor of mp4 sources. A fix is to set the sources when
ABR and HLS are not available.