🐛(encoding) respect original video aspect ratio
All our video presets for AWS MediaConvert, which control the way in
which videos are encoded to all the formats we need in Marsha, included
the "STRETH_TO_OUTPUT" option. This is not what we want (we'd rather have
black bars for eg. 4/3 videos) and was only there due to the files we
used as boilerplates to bootstrap our MediaConvert presets.

We also remove the "Width" parameter for video outputs. This way videos
with a different aspect ratio will not have the black bars hard-encoded
in them. We can add them in the player if needed, or display the video
as-is there.