💄(frontend) don't force aspect ratio on the player
We used to encode all video to 16/9 aspect ratio and, from the
assumtion that all videos would be in that format, set a base
aspect ratio in CSS for the player as 16/9, to avoid it having 0
height then moving around when the video is loaded.

We can change this to a min-height so video height is respected
if it's eg. 4/3. Videos wider than 16/9 will have black bars and
videos taller than 16/9 will be respected but "jump" once when the
video is loadeed and the player gets its final height.