✅(settings) add tests for the get_release method
As we ported the code that gets the release info in settings and
makes it available to another Django project, we noticed a possible
improvement: it was not tested.

By getting the release inside a function and just delegating to that
in the computed prop, we can add tests to make sure the code to get
release info works as intended.