🐛(back) set SECURE_REFERRER_POLICY setting to same-origin
Django 3.0 introduces a new setting `SECURE_REFERRER_POLICY` in its
`SecurityMiddleware`. There is no value by default but heartbeat requires
to set one in production. We only need the referer for the admin module
so we choose to use `same-origin`, this is the the value recommended by
the documentation to prevent CSRF issues:
https://docs.djangoproject.com/en/3.0/ref/middleware/#referrer-policy