🐛(front) force to display thumbnail while the video is not played
We have a regression on the thumbnail display feature. Before we forced
the frame code to 0 on the first seeked event because it happened before
the initialize event. We change how the initialize event is catch and it
can happen now before the first seeked event. To fix this we have to
check in the initialized event itself if we have to reset the
`currentTime` property.